Feature: included hexagonal particle arrays by Pawel Sobieszczyk. Reorganisation of create functions and removal of annoying unused variables in read functions.
