+++ src/create/create.cpp
@@ -132,7 +132,7 @@ int create(){
 	cs::create_crystal_structure(catom_array);
 
 	// Cut system to the correct type, species etc
-	cs::create_system_type(catom_array);
+	create::create_system_type(catom_array);
 
 	// Copy atoms for interprocessor communications
 	#ifdef MPICF
