+++ src/create/makefile
@@ -11,6 +11,7 @@ cs_set_atom_vars2.o \
 alloy.o \
 bubble.o \
 bulk.o \
+centre_particle.o \
 cone.o \
 compare_radius.o \
 composition.o \
@@ -20,13 +21,17 @@ data.o \
 ellipse.o \
 ellipsoid.o \
 faceted.o \
+hexagonal_array.o \
 initialize.o \
 interface.o \
 layers.o \
 multilayers.o \
 mpi.o \
+particle.o \
 roughness.o \
+sort_atoms_by_grain.o \
 sphere.o \
+square_array.o \
 teardrop.o \
 truncated_octahedron.o \
 voronoi.o \
