+++ src/vio/read.cpp
@@ -88,12 +88,11 @@ namespace vin{
 		zlog << zTs() << "Parsing system parameters from main input file." << std::endl;
 
 		int line_counter=0;
-                // Comment and delimiter characters for "input" type files. 
-                char com = '#';
-                char delim[] = ":=!";
+
+      // Comment and delimiter characters for "input" type files.
+      char delim[] = ":=!";
 
 		// Loop over all lines and pass keyword to matching function
-                std::vector<char> cstr[0];
 		while (! inputfile.eof() ){
 			line_counter++;
 
@@ -107,7 +106,7 @@ namespace vin{
 
 			// clear carriage return for dos formatted files
 			line.erase(remove(line.begin(), line.end(), '\r'), line.end());
-                        
+
 			// strip key,word,unit,value
 			std::string key="";
 			std::string word="";
@@ -115,28 +114,27 @@ namespace vin{
 			std::string unit="";
 
 			// get size of string
-			int linelength = line.length();
 
-                        // remove everything after comment character
-                        line = line.substr(0,line.find('#')) ;
+         // remove everything after comment character
+         line = line.substr(0,line.find('#')) ;
 
-                        // convert to c-string style, for tokenisation
-                        std::vector<char> cstr(line.begin(),line.end());
+         // convert to c-string style, for tokenisation
+         std::vector<char> cstr(line.begin(),line.end());
 
-                        // add null terminator, to insure that strtok cannot run over into previously used memory. 
-                        cstr.push_back('\0'); 
+         // add null terminator, to insure that strtok cannot run over into previously used memory.
+         cstr.push_back('\0');
 
-                        // tokenise the string, using delimiters from above
-                        char *token = strtok(&cstr[0],delim); // first call of strtok sets the string to tokenise. 
-                        for (int count = 0; count < 4 && token !=NULL; count++){
-                            if (count==0){key=token;}       // Format is always the same
-                            else if(count==1){word=token;}  // but breaks if EOL found
-                            else if(count==2){value=token;} // so if unused, keywords will remain as ""
-                            else if(count==3){unit=token;}
-                            token = strtok(NULL,delim);
-                            };
+         // tokenise the string, using delimiters from above
+         char *token = strtok(&cstr[0],delim); // first call of strtok sets the string to tokenise.
+         for (int count = 0; count < 4 && token !=NULL; count++){
+             if (count==0){key=token;}       // Format is always the same
+             else if(count==1){word=token;}  // but breaks if EOL found
+             else if(count==2){value=token;} // so if unused, keywords will remain as ""
+             else if(count==3){unit=token;}
+             token = strtok(NULL,delim);
+         };
 
-                        // tidy up
+         // tidy up
 			string empty="";
 			if(key!=empty){
 			//std::cout << "\t" << "key:  " << key << std::endl;
